video_llm_demo_offline/requirements.txt
# Core
requests==2.31.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydub==0.25.1

# Local STT (Whisper)
openai-whisper==20231117

# Image captioning (BLIP via transformers)
transformers==4.43.2
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0

# TTS (Kitty/Kitten TTS CLI)
kitty-tts==0.9.0
